Level-
added somethings. changed somethings. shader doesnt work. so need a newtoon shader that works with URP or LWRP. so ill be fixing that also making grass cause eli grass doesnt work. needs to be more specific. so the textures arent done because solution to shader is making new shader. ill see what i can do aka slap basic colours on shit and that be it.