Removed old Debug.Logs, commented out BaseSkills OnDrawGizmos
If you want to see the debug lines for the skills OnDrawGizmos, just enable it on your branch and disable it before merging back in. This causes a lot of errors to be logged because for some reason when an enemy dies the SimpleEnemy script is just removed from the object. ? I don't know why this has been done